Update support for FilterOptions (#10)
Updates support for FilterOptions -- the backend now expects a flat dictionary of string => string, encoded as JSON.

It also introduces some fixes for duplicate names. Since the philosophy of the SDK is to hide implementation detail of TargetGroup and SourceGroup, it deduplicates those internally if it finds a name that would be a duplicate.

Note that it doesn't yet support duplicate Alerts, but that might be a great next step if we find buggy behavior.