Move JWT storage to localforage (#17)
* Move to localforage

There are some clients for whom localStorage is not the right solution.

Migrate to localforge to abstract underlying storage one more step

* Catch errors inside hook

There's probably a better error handling pattern. However, a failure to
persist the JWT isn't the end of the world, so for now, just catch the
issues inside the hook.