Configurable Card library v0 (#22)
* Create new package

The package will be used to contain the React Card component.
Integrators are expected to supply their own modal containers.

* Define initial Folder structure

Put things into functional folders. The assumption is that users will be
depending on all of the components, transitively through a convenient
top-level component.

* Subscribe on hooks

The alert config is used to determine the alerts to create

* Add libphonenumber-js

This will be used to parse input strings etc.

* Remove style and copy contexts

This feels overwhelming, just pass as props

* Reduce re-render in useNotifiSubscribe

Don't depend on intermediary changes to update state

* Wrap a card around context

The card helps define intended usage around null signer and
walletPublicKey and enforces their non-nullness before rendering a
Context and the children.

While I was at it, I exposed NotifiEnvironment through and removed usage
of BlockchainEnvironment

* Fix Build
- Remove countryCode from SMS
- Export notifiCard
- Export hooks
- Update version number
- Fix SmsInput
- Fix rendered values

* Fetch data on mount

I think this belongs in the context but pull it here for now

* Add type to directpush

Allow service users to send a "type" along with directPush.

* Move AlertConfiguration to utils

Extract the type and add some factory functions
HealthThreshold is another utility method

* Update README with sample

Hopefully this is sufficient to outline some major things about the
library. Other examples might be useful once we get some feedback