Hardware support (#49)
* Add hardware support in useNotifiSubscribe

This is a pretty important feature for many NFT holders. Take on a
@solana/web3.js dependency in notifi-react-card to speed up development.

* Update subscribe to use `logIn`

The function now depends on `logIn` rather than `clientLogIn`

* Construct publicKey

Do not depend on the context

* Move solana dependencies to peer

This will prevent the library from packaging its own version (which
would likely result in conflicts with Context

* Move dependencies to props

I think there are some problems with the hooks when versions are
mismatched.