Aptos FrontendClient (#60)
* Bootstrap graphql package

This package will serve as the GQL dependency for other projects. Its
main purpose is to serve as a thin wrapper around the GQL api
(previously serviced by notifi-axios-adapter) but with better tooling
around GQL.

* Pull in a few Fragments

Get started with a few fragments to get the tooling warmed up

* Add LoginByTransaction operations

These are also a straightforward port

* Add the Create operations

Copy and paste more GQL, with some cleanups.

* Add Delete and Update mutations

Copy / pasted for now, but probably better to merge the `update`
mutations into the `create` onces in the near future.

* Add miscellaneous mutations

These are the remaining mutations from the previous package.

* Add queries from previous version

Port the queries to the new tooling.

* Update formatting rule

Import order was bothering me, so added the rule

* Bootstrap frontend-client repo

This repo will be used to contain business logic around interfacing with
the GraphQL package. The API will hopefully be re-usable across all
clients, regardless of blockchain or frontend framework.

* Add Request Headers to Service

Let the Service instance manage a current jwt and set request headers
for each request. It will also need a request ID.

* Create Rough outline for FrontendClient

The FrontendClient is intended to be the main API facade for integrators
of Notifi who are using it inside a web app, e.g. to allow users to
subscribe.

* Add completeLoginViaTransaction and beginLoginViaTransaction functions in Frontend Client Package (#56)

* refactor: add beginLoginViaTransaction to frontend client

* refactor: add completeLoginViaTransaction into frontendClient

* chore: update package-lock

* revert: codegen.yml

* chore: revert to original codegen.yml

* chore: correct type

* chore: throw new Error

* chore: clientRandomUuid as instance variable

* chore: resolve path to package

* chore: utilize types from notifi-graphql

* chore: do not track DS_Store

* Remove .DS_Store from everywhere

* chore: initialize clientRandomUuid to null

* chore: install notifi-network/notifi-graphql

* chore: utilize global object of crypto directly and initalize instance variable

* chore: explicit if not null

* chore: nonce as logValue

* chore: throw error if nonce null and clear out this._clientRandomUuid

Co-authored-by: Jennifer Tran <jennifertran@Jennifers-MacBook-Pro.local>

* Add tsconfig in notifi-frontend-client

This include seems to be required for the global npx run build to
succeed

* Implement ensureTargetGroup

This allows users to create or update a given target group

* Implement EnsureSourceGroup

This calls ensureSource in series under the hood.

* Move files around

Add index modules for all folders

* Add getSourceGroups

This is to expose the sourcegroups that have alreayd been creatd.

* Add createAlert / deleteAlert

This will finish the base use cases for the SDK v2

* Fork login method for walletBlockchain

For now, fork the method inside the client.

In the future, this should be delegated to some other wrapper, and the
client should use the wrapper as a driver.

* Make packages public

Without these, scoped packages default to private and paid.

* Update keyPrefix for storage driver

This code also depends on config

* Export storage

This is useful

* Add factory methods for AptosClient

Added Graphql-request dep to frontend-client and created factory methods

* Export from client

The two compile units were missing

* Add walletBlockchain input

This is necessary to resolve the right verification methods.

* Fix signing message

Move common string to constant

Co-authored-by: Jennifer Tran <jennifertran@Jennifers-MacBook-Pro.local>