Various fixes (#68)
* Add in-memory storage driver

In case local storage etc. are not available.

* Use in-memory storage in notifiStorage

We should parameterize this but for now use in-memory

* Conditionally instantiate rejected promise

Creating the rejected promise logs an error to the console. Tho it
wasn't preventing correct behavior, it was making Jest fail.

only instantiate it when there is a failure in order to fix this issue

* Remove telegramID from createDirectPushAlert

This is not exposed on GQL

* Add axios 0.27 to peer dependency

Allow peers to resolve to 0.27.2