TenantUser and Subscription (#154)
* Add getTenantUser API to Node SDK

This allows SID/Secret authenticated users to iterate across the full
set of users

* Add subscribe functionality

Rather than exposing it as another NotifiClient method, expose a
separate utility function. This helps to clarify the fact that the
subscription client should be disposed separately

The graphql-ws library is not yet supported by our server
Use the deprecated subscriptions-transport-ws library

* Add type assertion for ExecutionResult

Do some manual type introspection to verify that the result is a known
type before emitting. Note that this means not all `next` events will be
propagated to the `onNext` handler