Implement tradingPair renderer (#163)
* Implement functional changes for Trading Pair

Define a few new types in order to pipe through changes for the Trading
Pair type. For now, just sign up for a direct tenant message.

* Take a styling pass

Do a pass for the CSS to make the pair look similar to what the design
ought to be.

I wrapped the Select and the Input in a div in order to better control
the border look & feel

* Add parsing logic for Alert Row

We are encoding the alert information in the name... which is not great,
but better than nothing.

Parse out the information and show it on the Alert Row.