Migrate notifi-node to notifi-graphql
Using the generated GQL is much safer and easier to maintain in the long
run. In preparation for additional work on the node SDK, migrate the
dependencies to notifi-graphql.

There are some remaining issues -- namely the subscription work. In the
future we should consider migrating these as well into notifi-graphql.
For now, it's straightforward to isolate since these are the only ones
that depend on the subscription library.