Make copy configurable from top level (#204)
* Pass through copy from top-level

The AlertListPreview has some bits of copy that would be useful to be
able to change via configuration. Pass these in from the top level for
now.

(Better would be to make it configurable in the Admin Panel, perhaps)

* Make various pieces of copy configurable

For now, pull in the config from the frontend config. In the future, we
should plan on pulling these values from the cardConfig from AP as well.

* Remove headerText from PreviewList

The header text is superseded by the AlertBox now.