Update NotifiTooltip (#205)
* Update NotifiTooltip

The tooltip default style was kind of hard to see in darkMode. Update
some of the darkmode values.

I also noticed that the positioning was a little bit awkward. Added a
few variables to help deduplicate and changed the "arrow" to be purely
CSS.

* Apply suggested changes

This method perfectly lines up the arrow to the center of the info icon.

Updated the background color on light mode to be more visible