Refactor AlertDetails Markup
Before, we were rendering a full-sized card over the whole thing.

Instead, extract the state up to the SubscriptionCard and alter the
rendering to be conditional on this state.

This drastically improves the flexibility, because it results in
simpler, cleaner CSS and allows us to use the layout engine rather than
having to pixel-push to match the container size.