Remove Connect button when isConnectedElsewhere
Since it's a little confusing, prefer to remove the old connect button
when in the error state. This will limit the CTA to be one per wallet.

Also, render a "Verified" text block when the account is connected.

Update the copy to be Verify / Verify anyway / Verified, as per design
review.