Add support for disconnecting ConnectedWallet (#207)
* Add support for disconnecting ConnectedWallet

We have an error handling flow for ConnectWallet where the user is
prompted when the wallet they are trying to connect is already
associated with another Notifi account.

Implement this as an in-line error message and secondary action link.

* Remove Connect button when isConnectedElsewhere

Since it's a little confusing, prefer to remove the old connect button
when in the error state. This will limit the CTA to be one per wallet.

Also, render a "Verified" text block when the account is connected.

Update the copy to be Verify / Verify anyway / Verified, as per design
review.