Add a name when the id is parameterized
We want to use the same card in multiple pages on the dApp,
parameterized by some value, to provide notifications.

When we use the same alert name, the configurations will step on each
other, resulting in one page's settings overriding another page's.

If the ID is an input value, parameterize the alert name in order to
avoid this situation